# feat(split transform): add regex support for separator in split transform