Fixes #1881.

The input strings for separator are escaped initially to allow for literal strings using any regex metacharacters as mentioned by @bruceg [here](https://github.com/timberio/vector/issues/1881#issuecomment-649830946).