@psinghal20 2020-07-07
@bruceg I have made the changes you mentioned. Also, I was trying to run the benchmarks using `cargo bench` but it gets stuck while running the `pipe` benchmark on warming-up stage.