@psinghal20 2020-07-09
Hi @bruceg, I ran the benchmarks for the changes. The results can be found [here](https://gist.github.com/psinghal20/311de314d8094bb0b00e4dce817f712f). I think there is a performance hit in some of the benchmarks, though it is not that clear. 

PS: There was some problem while running the normal benchmarks for me, all the benching process were just going into sleep and waiting for some event. Don't know why. I ran the wasm benchmarks for the above results.