@bruceg 2020-07-11
I wasn't clear. There is currently no benchmark to test this. My comment was that we should make a benchmark to show this doesn't cause a performance regression. I'll write something up.