@psinghal20 2020-07-12
Oh sorry! I thought you mean the current benchmark numbers. :sweat_smile: 

I will try writing a benchmark similar to the others for split transform. Meanwhile, I wrote a simple benchmark just for the Regex vs String split functions. 
```
Split/Without regex     time:   [64.972 ns 65.680 ns 66.498 ns]
                        change: [+0.3620% +1.9603% +3.5347%] (p = 0.01 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  1 (1.00%) high mild
  7 (7.00%) high severe
Split/With regex        time:   [90.120 ns 90.973 ns 91.834 ns]
                        change: [-3.9364% -2.4538% -0.9200%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe
```
This benchmark doesn't count the regex compilation time and uses the same input string. Considering this, I think it might introduce some performance overhead. I will try to write the benchmark for the split transform tomorrow.