@psinghal20 2020-07-16
Hi @bruceg, I wrote a benchmark for the split transform. Here are the results:

Without Regex 
```
split_transform         time:   [939.83 ns 964.64 ns 990.52 ns]
Found 8 outliers among 100 measurements (8.00%)
  8 (8.00%) high mild
```

With Regex
```
split_transform         time:   [1.4710 us 1.5069 us 1.5491 us]
                        change: [+64.450% +69.399% +74.054%] (p = 0.00 < 0.05)
                        Performance has regressed.
```

You can find the pdf plot for the same here: https://gist.githubusercontent.com/psinghal20/680ae1d0657ae8f11b7381132510a027/raw/34cf4620e882ddf507c1c72f69594dde90c531aa/pdf.svg

I believe this change does introduce performance overhead and we might look into keeping this config literal till we don't require regex explicitly.

@Hoverbear thanks for your help!